refactor: Refactor all AI-NLQ related scipts
